requirements.txt
openai
python-dotenv
pyaudio
sounddevice
soundcard
numpy
customtkinter
soundfile 